% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamant.R
\name{adamant}
\alias{adamant}
\title{Adaptive Mantel test for linear association.}
\usage{
adamant(X, Y, lambdas_X = c(0), n_perms = 100, verbose = TRUE,
  P_val_only = FALSE)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{Y}{A vector of response values.}
}
\value{
An object of class `adamant` that contains the test results.
}
\description{
\code{adamant} Returns a dataframe containing results of the adaptive
Mantel test for linear association between two sets of variables.
}
\examples{
\dontrun{
X <- matrix(rnorm(100), nrow = 20, ncol = 5)
Y <- rnorm(20)
lambdas_X <- c(0, 1, 1000, Inf)
adamant(X, Y, lambdas_X)
}
}
